change CoreTask.courseId from int to long

+++ src/co/uberdev/ultimateorganizer/core/CoreTask.java
@@ -39,7 +39,7 @@ public class CoreTask extends Core
 	// The related course of the task
 	protected CoreCourse course;
 	// the id on the server side of this course
-	protected int courseId;
+	protected long courseId;
 	// The code of this course
 	protected String courseCodeCombined;
     // dateCreated signifies the creation date of the task, will be initialised when the task is created
@@ -292,11 +292,11 @@ public class CoreTask extends Core
         return taskOwnerNameCombined;
     }
 
-	public int getCourseId() {
+	public long getCourseId() {
 		return courseId;
 	}
 
-	public void setCourseId(int courseId) {
+	public void setCourseId(long courseId) {
 		this.courseId = courseId;
 	}
